+++ src/XHS.Spider/Views/Windows/ScanLogin.xaml
@@ -7,7 +7,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:XHS.Spider.Views.Windows"
         mc:Ignorable="d"
-        Title="扫码登录" Height="280" Width="210">
+        Title="扫码登录" Height="300" Width="240">
     <Grid Margin="0,0,0,3">
         <Image
             Name="nameLoginQRCode"
