+++ src/XHS.Spider/Views/Windows/ScanLogin.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,6 +15,10 @@ using System.Windows.Shapes;
 using Wpf.Ui.Controls;
 using Wpf.Ui.Controls.Interfaces;
 using Wpf.Ui.Mvvm.Contracts;
+using XHS.Common.Helpers;
+using XHS.IService.XHS;
+using XHS.Service.Log;
+using XHS.Service.XHS;
 using XHS.Spider.ViewModels;
 
 namespace XHS.Spider.Views.Windows
@@ -23,17 +28,60 @@ namespace XHS.Spider.Views.Windows
     /// </summary>
     public partial class ScanLogin : Window, INavigationWindow
     {
-        public ScanLoginViewModel ViewModel
+        private static readonly Service.Log.ILogger Logger = LoggerService.Get(typeof(ScanLogin));
+        #region 属性
+        private readonly IXhsSpiderService _xhsSpiderService;
+        #endregion
+        public ScanLogin(IXhsSpiderService xhsSpiderService)
         {
-            get;
-        }
-        public ScanLogin(ScanLoginViewModel viewModel)
-        {
+            _xhsSpiderService = xhsSpiderService;
             WindowStartupLocation = WindowStartupLocation.CenterScreen;// 窗体居中
-            this.ViewModel= viewModel;
             InitializeComponent();
+            InitQrCode();
         }
+        public async void InitQrCode()
+        {
+            try
+            {
+                var qrcodeInfo = await _xhsSpiderService.CreateQrCode();
+                if (qrcodeInfo != null)
+                {
+                    var qrcode = GetLoginQRCode(qrcodeInfo.Url);
+                    if (qrcode != null)
+                    {
+                        App.PropertyChangeAsync(new Action(async () =>
+                        {
+                            this.nameLoginQRCode.Source = qrcode;
+                        }));
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                //Wpf.Ui.Controls.MessageBox.("二维码获取失败");
+                Logger.Error("二维码获取失败", ex);
+            }
+        }
+        /// <summary>
+        /// 根据输入url生成二维码
+        /// </summary>
+        /// <param name="url"></param>
+        /// <returns></returns>
+        public static BitmapImage GetLoginQRCode(string url)
+        {
+            // 设置的参数影响app能否成功扫码
+            System.Drawing.Bitmap qrCode = QRCode.EncodeQRCode(url, 10, 10, null, 0, 0, false);
 
+            MemoryStream ms = new MemoryStream();
+            qrCode.Save(ms, System.Drawing.Imaging.ImageFormat.Bmp);
+            byte[] bytes = ms.GetBuffer();
+            ms.Close();
+            BitmapImage image = new BitmapImage();
+            image.BeginInit();
+            image.StreamSource = new MemoryStream(bytes);
+            image.EndInit();
+            return image;
+        }
         public void CloseWindow()
         {
             throw new NotImplementedException();
